Ajout d'une base pour les évenements de week end (rotation entre les events, calendrier) (#1221)
* rewrite contest system launch + add WeeklyEvent for more flexibility

* Rewrite contest schedule + gestion for multiple events

* fix init

* add CalendarMenu and /events
