+++ src/main/java/fr/openmc/core/CommandsManager.java
@@ -11,7 +11,8 @@ import fr.openmc.core.features.adminshop.AdminShopCommand;
 import fr.openmc.core.features.animations.DebugAnimationCommand;
 import fr.openmc.core.features.credits.CreditsCommand;
 import fr.openmc.core.features.cube.CubeCommands;
-import fr.openmc.core.features.events.halloween.commands.HalloweenCommands;
+import fr.openmc.core.features.events.commands.EventCommand;
+import fr.openmc.core.features.events.contents.halloween.commands.HalloweenCommands;
 import fr.openmc.core.features.friend.FriendCommand;
 import fr.openmc.core.features.mailboxes.MailboxCommand;
 import fr.openmc.core.features.mainmenu.commands.MainMenuCommand;
@@ -61,7 +62,8 @@ public class CommandsManager {
                 new CreditsCommand(),
                 new CubeCommands(),
                 new HalloweenCommands(),
-                new DebugAnimationCommand()
+                new DebugAnimationCommand(),
+                new EventCommand()
         );
     }
 }
