+++ src/main/java/fr/openmc/core/features/events/contents/weeklyevents/contents/contest/commands/autocomplete/ColorContestAutoComplete.java
@@ -1,0 +1,17 @@
+package fr.openmc.core.features.events.contents.weeklyevents.contents.contest.commands.autocomplete;
+
+import fr.openmc.core.features.events.contents.weeklyevents.contents.contest.managers.ContestManager;
+import org.jetbrains.annotations.NotNull;
+import revxrsal.commands.autocomplete.SuggestionProvider;
+import revxrsal.commands.bukkit.actor.BukkitCommandActor;
+import revxrsal.commands.node.ExecutionContext;
+
+import java.util.List;
+
+public class ColorContestAutoComplete implements SuggestionProvider<BukkitCommandActor> {
+
+    @Override
+    public @NotNull List<String> getSuggestions(@NotNull ExecutionContext<BukkitCommandActor> context) {
+        return ContestManager.getColorContestList();
+    }
+}
