+++ src/main/java/fr/openmc/core/features/leaderboards/LeaderboardManager.java
@@ -7,8 +7,8 @@ import fr.openmc.core.features.city.CityManager;
 import fr.openmc.core.features.economy.BankManager;
 import fr.openmc.core.features.economy.EconomyManager;
 import fr.openmc.core.features.economy.models.EconomyPlayer;
-import fr.openmc.core.features.events.halloween.managers.HalloweenManager;
-import fr.openmc.core.features.events.halloween.models.HalloweenData;
+import fr.openmc.core.features.events.contents.halloween.managers.HalloweenManager;
+import fr.openmc.core.features.events.contents.halloween.models.HalloweenData;
 import fr.openmc.core.features.leaderboards.commands.LeaderboardCommands;
 import fr.openmc.core.utils.DateUtils;
 import fr.openmc.core.utils.entities.TextDisplay;
