+++ src/main/java/fr/openmc/core/features/quests/quests/WinContestQuest.java
@@ -1,6 +1,6 @@
 package fr.openmc.core.features.quests.quests;
 
-import fr.openmc.core.features.contest.ContestEndEvent;
+import fr.openmc.core.features.events.contents.weeklyevents.contents.contest.events.ContestEndEvent;
 import fr.openmc.core.features.quests.objects.Quest;
 import fr.openmc.core.features.quests.objects.QuestTier;
 import fr.openmc.core.features.quests.rewards.QuestItemReward;
