+++ src/main/java/fr/openmc/core/utils/database/DatabaseManager.java
@@ -12,12 +12,13 @@ import fr.openmc.core.features.city.sub.notation.NotationManager;
 import fr.openmc.core.features.city.sub.rank.CityRankManager;
 import fr.openmc.core.features.city.sub.statistics.CityStatisticsManager;
 import fr.openmc.core.features.city.sub.war.WarManager;
-import fr.openmc.core.features.contest.managers.ContestManager;
 import fr.openmc.core.features.dream.DreamManager;
 import fr.openmc.core.features.economy.BankManager;
 import fr.openmc.core.features.economy.EconomyManager;
 import fr.openmc.core.features.economy.TransactionsManager;
-import fr.openmc.core.features.events.halloween.managers.HalloweenManager;
+import fr.openmc.core.features.events.contents.halloween.managers.HalloweenManager;
+import fr.openmc.core.features.events.contents.weeklyevents.WeeklyEventsManager;
+import fr.openmc.core.features.events.contents.weeklyevents.contents.contest.managers.ContestManager;
 import fr.openmc.core.features.friend.FriendSQLManager;
 import fr.openmc.core.features.homes.HomesManager;
 import fr.openmc.core.features.mailboxes.MailboxManager;
@@ -63,6 +64,7 @@ public class DatabaseManager {
             AnalyticsManager.initDB(connectionSource);
             MailboxManager.initDB(connectionSource);
             ContestManager.initDB(connectionSource);
+            WeeklyEventsManager.initDB(connectionSource);
             EconomyManager.initDB(connectionSource);
             HomesManager.initDB(connectionSource);
             FriendSQLManager.initDB(connectionSource);
